Use `State` property in the sample app
Use a `State` property in a presenter model in the sample app. Updates of this value are not propagated through a new [Model] and consumers must observe the value directly instead. In a Compose UI layer this is trivial, because the Compose runtime does this automatically.

Using this approach for the timeout is much more efficient, because the value changes every 10 milliseconds and going through the whole presenter tree to compute a new model and updating all renderers adds a lot of load. With this approach only the necessary Composable UI element is updated.