Add a recipe for the Navigation 3 library
The Navigation 3 library is in early alpha stages. This change demonstrates how Navigation 3 can be used with App Platform. A new screen in the Recipes app manages its own backstack and forwards navigation events from Navigation 3 to the presenter.

Navigation 3 is only available for Android for now. On other platforms a placeholder UI is shown.