Fix and stop suppressing NPE when removing Android Views
In the past we were seeing this crash without being able to reproduce it:
```
 java.lang.IndexOutOfBoundsException
 	at android.view.ViewGroup.removeFromArray(ViewGroup.java:5665)
 	at android.view.ViewGroup.removeViewInternal(ViewGroup.java:5853)
 	at android.view.ViewGroup.removeViewInternal(ViewGroup.java:5815)
 	at android.view.ViewGroup.removeView(ViewGroup.java:5746)
 	at software.amazon.app.platform.renderer.ViewRenderer.resetView(ViewRenderer.kt:100)
 	at software.amazon.app.platform.renderer.ViewRenderer.access$resetView(ViewRenderer.kt:63)
 	at software.amazon.app.platform.renderer.ViewRenderer$render$lambda$4$$inlined$doOnDetach$1.onViewDetachedFromWindow(View.kt:436)
 	at android.view.View.dispatchDetachedFromWindow(View.java:23314)
```
Because of that we simply caught and ignored the exception, but this caused more issues later at runtime.

We managed to reproduce this particular bug now and addressed the NPE properly. It no longer needs to be suppressed.

Fixes #136