Fix potential error in renderer documentation
The current documentation educates the reader that Compose Multiplatform and Android Views are supported out of the box by the 'Renderer'. However, this is contradicted by saying that Compose Multiplatform requires an opt-in in the next line.
The commit changes this to Compose UI, in line with the Gradle DSL config `enableComposeUi`.